linux (Generic): 2nd kernel 4.19 backported from 9.0
there are some stutter issues with MPEG2 in 9.2, so i did take
the 4.19 kernel from 9.0 and updated it to the latest 4.19.71.
kernel config is copied from 9.0 too as are the patches.
to build you do "LINUX=generic-4.19 make image"

linux (Generic): for generic-4.19 linux-4.19.72
linux (Generic): for generic-4.19 linux-4.19.74
linux (Generic): for generic-4.19 linux-4.19.75
linux (Generic): for generic-4.19 linux-4.19.76