busybox: init: fix toram when SYSTEM_IMAGE is with path
For the SYSTEM copy to /dev always use /dev/SYSTEM as target name
to not deal with fancy path or filenames.
Thanks HiassofT for clearing it up for me why that is best.
As far as i can tell all the rest of init is fine with BOOT_IMAGE and
SYSTEM_IMAGE having slashes in there.
Just toram was broken.